requirements.txt
python-nmap
openai
python-dotenv
requests